requirements.txt
streamlit
torch
torchvision
opencv-python-headless
albumentations
segmentation-models-pytorch
pandas
numpy
Pillow
timm
scikit-learn
tqdm
matplotlib
